# openBanking/src/main/webapp/META-INF/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: ServiceDesk@openbanking.org.uk
    name: Service Desk
  description: Swagger for Payment Initiation API Specification
  license:
    name: open-licence
    url: https://www.openbanking.org.uk/open-licence
  termsOfService: https://www.openbanking.org.uk/terms
  title: Payment Initiation API
  version: 3.1.10
servers:
- url: /
- url: /open-banking/v3.1/pisp
paths:
  /domestic-payment-consents:
    post:
      operationId: CreateDomesticPaymentConsents
      parameters:
      - description: "The time when the PSU last logged in with the TPP. \nAll dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: \nSun, 10 Sep 2017 19:43:31 UTC"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\
            \ \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$"
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: |
          Every request will be processed only once per x-idempotency-key.  The
          Idempotency Key will be valid for 24 hours.
        explode: false
        in: header
        name: x-idempotency-key
        required: true
        schema:
          maxLength: 40
          pattern: ^(?!\s)(.*)(\S)$
          type: string
        style: simple
      - description: A detached JWS signature of the body of the payload.
        explode: false
        in: header
        name: x-jws-signature
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates the user-agent that the PSU is using.
        explode: false
        in: header
        name: x-customer-user-agent
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/OBWriteDomesticConsent4'
          application/json:
            schema:
              $ref: '#/components/schemas/OBWriteDomesticConsent4'
          application/jose+jwe:
            schema:
              $ref: '#/components/schemas/OBWriteDomesticConsent4'
        description: Default
        required: true
      responses:
        "201":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
            application/json:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
          description: Domestic Payment Consents Created
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Bad request
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          description: Unauthorized
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Forbidden
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "405":
          description: Method Not Allowed
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "406":
          description: Not Acceptable
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "415":
          description: Unsupported Media Type
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Number in seconds to wait
              explode: false
              schema:
                type: integer
              style: simple
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Internal Server Error
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
      security:
      - TPPOAuth2Security:
        - payments
      summary: Create Domestic Payment Consents
      tags:
      - Domestic Payments
      x-contentType: application/json; charset=utf-8
      x-accepts: application/json
      x-tags:
      - tag: Domestic Payments
  /domestic-payment-consents/{ConsentId}:
    get:
      operationId: GetDomesticPaymentConsentsConsentId
      parameters:
      - description: ConsentId
        explode: false
        in: path
        name: ConsentId
        required: true
        schema:
          type: string
        style: simple
      - description: "The time when the PSU last logged in with the TPP. \nAll dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: \nSun, 10 Sep 2017 19:43:31 UTC"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\
            \ \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$"
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates the user-agent that the PSU is using.
        explode: false
        in: header
        name: x-customer-user-agent
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
            application/json:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBWriteDomesticConsentResponse5'
          description: Domestic Payment Consents Read
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Bad request
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          description: Unauthorized
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Forbidden
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "405":
          description: Method Not Allowed
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "406":
          description: Not Acceptable
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Number in seconds to wait
              explode: false
              schema:
                type: integer
              style: simple
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Internal Server Error
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
      security:
      - TPPOAuth2Security:
        - payments
      summary: Get Domestic Payment Consents
      tags:
      - Domestic Payments
      x-accepts: application/json
      x-tags:
      - tag: Domestic Payments
  /domestic-payment-consents/{ConsentId}/funds-confirmation:
    get:
      operationId: GetDomesticPaymentConsentsConsentIdFundsConfirmation
      parameters:
      - description: ConsentId
        explode: false
        in: path
        name: ConsentId
        required: true
        schema:
          type: string
        style: simple
      - description: "The time when the PSU last logged in with the TPP. \nAll dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: \nSun, 10 Sep 2017 19:43:31 UTC"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\
            \ \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$"
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates the user-agent that the PSU is using.
        explode: false
        in: header
        name: x-customer-user-agent
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBWriteFundsConfirmationResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBWriteFundsConfirmationResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBWriteFundsConfirmationResponse1'
          description: Domestic Payment Consents Read
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Bad request
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          description: Unauthorized
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Forbidden
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "405":
          description: Method Not Allowed
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "406":
          description: Not Acceptable
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Number in seconds to wait
              explode: false
              schema:
                type: integer
              style: simple
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/json:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
            application/jose+jwe:
              schema:
                $ref: '#/components/schemas/OBErrorResponse1'
          description: Internal Server Error
          headers:
            x-fapi-interaction-id:
              description: An RFC4122 UID used as a correlation id.
              explode: false
              required: true
              schema:
                type: string
              style: simple
            x-jws-signature:
              description: |
                Header containing a detached JWS signature of the body of the payload.
              explode: false
              required: true
              schema:
                type: string
              style: simple
      security:
      - PSUOAuth2Security:
        - payments
      summary: Get Domestic Payment Consents Funds Confirmation
      tags:
      - Domestic Payments
      x-accepts: application/json
      x-tags:
      - tag: Domestic Payments
...
components:
...
  schemas:
...
    OBErrorResponse1:
      additionalProperties: false
      description: "An array of detail error codes, and messages, and URLs to documentation\
        \ to help remediation."
      properties:
        Code:
          description: "High level textual error code, to help categorize the errors."
          maxLength: 40
          minLength: 1
          type: string
        Id:
          description: "A unique reference for the error instance, for audit purposes,\
            \ in case of unknown/unclassified errors."
          maxLength: 40
          minLength: 1
          type: string
        Message:
          description: "Brief Error message, e.g., 'There is something wrong with\
            \ the request parameters provided'"
          maxLength: 500
          minLength: 1
          type: string
        Errors:
          items:
            $ref: '#/components/schemas/OBError1'
          minItems: 1
          type: array
      required:
      - Code
      - Errors
      - Message
      type: object
...
    OBWriteDomesticConsent4:
      additionalProperties: false
      example:
        Risk:
          PaymentPurposeCode: PaymentPurposeCode
          BeneficiaryPrepopulatedIndicator: true
          ContractPresentInidicator: true
          PaymentContextCode: BillingGoodsAndServicesInAdvance
          DeliveryAddress:
            StreetName: StreetName
            CountrySubDivision: CountrySubDivision
            AddressLine:
            - AddressLine
            - AddressLine
            BuildingNumber: BuildingNumber
            TownName: TownName
            Country: Country
            PostCode: PostCode
          MerchantCategoryCode: MerchantCategoryCode
          MerchantCustomerIdentification: MerchantCustomerIdentification
          BeneficiaryAccountType: null
        Data:
          SCASupportData:
            RequestedSCAExemptionType: BillPayment
            AppliedAuthenticationApproach: CA
            ReferencePaymentOrderId: ReferencePaymentOrderId
          Authorisation:
            CompletionDateTime: 2000-01-23T04:56:07.000+00:00
            AuthorisationType: Any
          ReadRefundAccount: "No"
          Initiation:
            SupplementaryData:
              key: "{}"
            LocalInstrument: null
            DebtorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            RemittanceInformation:
              Unstructured: Unstructured
              Reference: Reference
            EndToEndIdentification: EndToEndIdentification
            InstructionIdentification: InstructionIdentification
            CreditorAccount:
              SecondaryIdentification: SecondaryIdentification
              SchemeName: null
              Identification: Identification
              Name: Name
            InstructedAmount:
              Amount: Amount
              Currency: Currency
            CreditorPostalAddress:
              StreetName: StreetName
              CountrySubDivision: CountrySubDivision
              Department: Department
              AddressLine:
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              - AddressLine
              BuildingNumber: BuildingNumber
              TownName: TownName
              Country: Country
              SubDepartment: SubDepartment
              AddressType: null
              PostCode: PostCode
      properties:
        Data:
          $ref: '#/components/schemas/OBWriteDomesticConsent4_Data'
        Risk:
          $ref: '#/components/schemas/OBRisk1'
      required:
      - Data
      - Risk
      type: object
...
  securitySchemes:
    TPPOAuth2Security:
      description: TPP client credential authorisation flow with the ASPSP
      flows:
        clientCredentials:
          scopes:
            payments: Generic payment scope
          tokenUrl: https://authserver.example/token
      type: oauth2
    PSUOAuth2Security:
      description: "OAuth flow, it is required when the PSU needs to perform SCA with\
        \ the ASPSP when a TPP wants to access an ASPSP resource owned by the PSU"
      flows:
        authorizationCode:
          authorizationUrl: https://authserver.example/authorization
          scopes:
            payments: Generic payment scope
          tokenUrl: https://authserver.example/token
      type: oauth2